+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-23 21:10:20
-// Commit: docs(core): update readme and API contracts for core
+// Updated at: 2026-04-24 15:15:36
+// Commit: style(core): harmonize light and dark theme palette for core
 
 class coreComponentState {
-  final String commitRef = '616';
-  final String lastMessage = 'docs(core): update readme and API contracts for core';
-  final DateTime timestamp = DateTime.parse('2026-04-23T21:10:20+05:00');
+  final String commitRef = '636';
+  final String lastMessage = 'style(core): harmonize light and dark theme palette for core';
+  final DateTime timestamp = DateTime.parse('2026-04-24T15:15:36+05:00');
 }